requirements.txt
django==5.1.3 
psycopg2-binary==2.9.9 
python-dotenv==1.0.1 
djangorestframework==3.15.2 
django-allauth==0.63.6 
bcrypt==4.2.0 
gunicorn==23.0.0